script: add fetching script
This script downloads web resources that are used in this repository.
Currently that is only Sean Corfield's dev.clj.

An alternative would be to use git submodules and symlinks, but I think
that is too much at this point. If I find myself using multiple
repositories that are frequently updated, I may go that route.